requirements.txt
streamlit
google-generativeai
cv2
os
numpy
PIL
# Secrets
.env